fix rubcop issues
